introduce --skip-tables, and allow diff to work with multiple connections

+++ src/Parse/CreateDatabase.php
@@ -132,6 +132,7 @@ class CreateDatabase
      * 'createTable' | (bool) include 'CREATE TABLE' statements [default: true]
      * 'dropTable'   | (bool) include 'DROP TABLE' statements [default: true]
      * 'alterEngine' | (bool) include ALTER TABLE ... ENGINE= [default: true]
+     * 'skipTables'  | string - regex of tables to ignore
      *
      * @return string[]
      */
@@ -141,6 +142,7 @@ class CreateDatabase
             'createTable' => true,
             'dropTable'   => true,
             'alterEngine' => true,
+            'skipTables'  => null,
         ];
 
         $thisTableNames = array_keys($this->tables);
@@ -152,26 +154,34 @@ class CreateDatabase
 
         $diff = [];
 
+        $skipTablesRegex = $flags['skipTables'];
+
         if ($flags['dropTable'] && count($droppedTableNames) > 0) {
             foreach($droppedTableNames as $tableName) {
-                $diff[] = "DROP TABLE IF EXISTS " . Token::escapeIdentifier($tableName);
+                if (!preg_match($skipTablesRegex, $tableName)) {
+                    $diff[] = "DROP TABLE IF EXISTS " . Token::escapeIdentifier($tableName);
+                }
             }
         }
 
         if ($flags['createTable'] && count($createdTableNames) > 0) {
             foreach($createdTableNames as $tableName) {
-                $diff = array_merge($diff, $that->tables[$tableName]->getDDL());
+                if (!preg_match($skipTablesRegex, $tableName)) {
+                    $diff = array_merge($diff, $that->tables[$tableName]->getDDL());
+                }
             }
         }
 
         foreach($commonTableNames as $tableName) {
-            $thisTable = $this->tables[$tableName];
-            $thatTable = $that->tables[$tableName];
-            $tableDiff = $thisTable->diff($thatTable, [
-                'alterEngine' => $flags['alterEngine'],
-            ]);
-            if ($tableDiff) {
-                $diff = array_merge($diff, $tableDiff);
+            if (!preg_match($skipTablesRegex, $tableName)) {
+                $thisTable = $this->tables[$tableName];
+                $thatTable = $that->tables[$tableName];
+                $tableDiff = $thisTable->diff($thatTable, [
+                    'alterEngine' => $flags['alterEngine'],
+                ]);
+                if ($tableDiff) {
+                    $diff = array_merge($diff, $tableDiff);
+                }
             }
         }
